# app/static/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "description": "Code With Prince",
      "version": "4.11.1",
      "title": "Code With Prince velaiapp"
     
    },
...
      "servers": [
        {
        "url": "/"
        },{
            "url":"/http"
        }
    ],
...
    "components": {
        "schemas": {
            "TodoList": {
                "type": "object",
                "properties": {
                    "number": {
                    "type": "string"
                    }
                   
                }
            },
...
        }
    },
...
    "paths": {
...
                "/api/shorttime_job": {
                    "post": {
                    "tags": [
                        "job_provider_post api"
                    ],
                    "summary": "Add shorttime job",
                    "requestBody": {
                    "description": "Post New todo To Database",
                    "required": true,
                    "content": {
                        "application/json": {
                        "schema": {
                            "$ref": "#/components5/schemas/TodoList"
                            }
                        }
                    }
                    },
                    "responses": {
                        "200": {
                            "description": "OK",
                            "schema": {
                            "$ref": "#/components/schemas/TodoList"
                            }
                        },
                        "404": {
                            "description": "NOT FOUND",
                            "schema": {
                            "$ref": "#/components/schemas/TodoList"
                            }
                        }
                    }

                }
                },
                    "/api/long_job_post": {
                        "post": {
                        "tags": [
                            "job_provider_post api"
                        ],
                        "summary": "Add longtime_job",
                        "requestBody": {
                        "description": "Post New todo To Database",
                        "required": true,
                        "content": {
                            
                            "application/json": {
                            "schema": {
                                "$ref": "#/components6/schemas/TodoList"
                                }
                            }
                        }
                        },
                        "responses": {
                            "200": {
                                "description": "OK",
                                "schema": {
                                "$ref": "#/components/schemas/TodoList"
                                }
                            },
                            "404": {
                                "description": "NOT FOUND",
                                "schema": {
                                "$ref": "#/components/schemas/TodoList"
                                }
                            }
                        }

                    }
                    
                    
                },
...
                            "/api/job_pro_userinfo_details": {
                                "post": {
                                "tags": [
                                    "job_provider_post api"
                                ],
                                "summary": "create_provider_info",
                                "requestBody": {
                                "description": "Post New todo To Database",
                                "required": true,
                                "content": {
                                    
                                    "application/json": {
                                    "schema": {
                                        "$ref": "#/components60/schemas/TodoList"
                                        }
                                    }
                                }
                                },
                                "responses": {
                                    "200": {
                                        "description": "OK",
                                        "schema": {
                                        "$ref": "#/components/schemas/TodoList"
                                        }
                                    },
                                    "404": {
                                        "description": "NOT FOUND",
                                        "schema": {
                                        "$ref": "#/components/schemas/TodoList"
                                        }
                                    }
                                }
            
                            }
                            
                            
                        },
...
    "/api/rental_pro_info_get": {
        "post": {
        "tags": [
            " Rental post API"
        ],
        "summary": "rental provider info ",
        "requestBody": {
        "description": "Post New todo To Database",

        "required": true,
       
        "content": {
            
            "application/json": {
            "schema": {
                "$ref": "#/rentalsee2/schemas/TodoList"
                }
            }
        }
        },
        "responses": {
            "200": {
                "description": "OK",
                "schema": {
                "$ref": "#/components/schemas/TodoList"
                }
            },
            "404": {
                "description": "NOT FOUND",
                "schema": {
                "$ref": "#/components/schemas/TodoList"
                }
            }
        }

    }
    
    
    },
    "/api/rental_pro_product_post": {
        "post": {
        "tags": [
            " Rental post API"
        ],
        "summary": "rental provider info ",
        "requestBody": {
        "description": "Post New todo To Database",

        "required": true,
    
        "content": {
            
            "application/json": {
            "schema": {
                "$ref": "#/rentalsee4/schemas/TodoList"
                }
            }
        }
        },
        "responses": {
            "200": {
                "description": "OK",
                "schema": {
                "$ref": "#/components/schemas/TodoList"
                }
            },
            "404": {
                "description": "NOT FOUND",
                "schema": {
                "$ref": "#/components/schemas/TodoList"
                }
            }
        }

    }


    },"/api/rented_products": {
        "post": {
        "tags": [
            " Rental post API"
        ],
        "summary": "rental provider info ",
        "requestBody": {
        "description": "Post New todo To Database",

        "required": true,
    
        "content": {
            
            "application/json": {
            "schema": {
                "$ref": "#/rentalsee5/schemas/TodoList"
                }
            }
        }
        },
        "responses": {
            "200": {
                "description": "OK",
                "schema": {
                "$ref": "#/components/schemas/TodoList"
                }
            },
            "404": {
                "description": "NOT FOUND",
                "schema": {
                "$ref": "#/components/schemas/TodoList"
                }
            }
        }

        }


    },
...
                "/api/rent_pro_post_show/{user_id}":{

                    "get" :{
                        "tags": [
                            "rental Get API"
                        ],
                        "summary": "Get ",
                        "parameters": [{
                        "name": "user_id",
                        "in": "path",
                        "description": "job_apply_count",
                        "required": true,
                        "type": "string",
                        "schema" : {
                            "$ref": "#/rentalsee2/schemas/user_id"
                        }
                        }],
                    
                        "responses": {
                        "200": {
                            "description": "OK",
                            "schema": {
                            "$ref": "#/components/schemas/TodoList"
                            }
                        },
                        "404": {
                            "description": "NOT FOUND",
                            "schema": {
                            "$ref": "#/components/schemas/TodoList"
                            }
                        }
                    }
                    }
                },
                "/api/rented_products/{user_id}":{

                    "get" :{
                        "tags": [
                            "rental Get API"
                        ],
                        "summary": "Get ",
                        "parameters": [{
                        "name": "user_id",
                        "in": "path",
                        "description": "job_apply_count",
                        "required": true,
                        "type": "string",
                        "schema" : {
                            "$ref": "#/rentalsee2/schemas/user_id"
                        }
                        }],
                    
                        "responses": {
                        "200": {
                            "description": "OK",
                            "schema": {
                            "$ref": "#/components/schemas/TodoList"
                            }
                        },
                        "404": {
                            "description": "NOT FOUND",
                            "schema": {
                            "$ref": "#/components/schemas/TodoList"
                            }
                        }
                    }
                    }
                },
                "/api/rent_pro_products/{id}":{

                    "get" :{
                        "tags": [
                            "rental Get API"
                        ],
                        "summary": "Get ",
                        "parameters": [{
                        "name": "id",
                        "in": "path",
                        "description": "job_apply_count",
                        "required": true,
                        "type": "string",
                        "schema" : {
                            "$ref": "#/rentalsee3/schemas/id"
                        }
                        }],
                    
                        "responses": {
                        "200": {
                            "description": "OK",
                            "schema": {
                            "$ref": "#/components/schemas/TodoList"
                            }
                        },
                        "404": {
                            "description": "NOT FOUND",
                            "schema": {
                            "$ref": "#/components/schemas/TodoList"
                            }
                        }
                    }
                    }
                },
...
    }
}